fix(keypatherror.ts): rename
